requirements.txt
PyPDF2>=3.0.0
pdfminer.six>=20221105
langchain>=0.0.340 